backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
rapidfuzz==3.9.3
spacy==3.8.11
openai==1.30.1
twilio==9.1.0
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
